+++ src/components/FormPanel/FormPanel.tsx
@@ -5,6 +5,7 @@ import {
   AppEvents,
   DataFrame,
   DataQueryError,
+  DataQueryResponse,
   Field,
   LoadingState,
   PanelProps,
@@ -31,6 +32,7 @@ import {
 import React, { useCallback, useEffect, useMemo, useState } from 'react';
 
 import {
+  ConfirmationElementDisplayMode,
   ContentType,
   FormElementType,
   LayoutOrientation,
@@ -52,6 +54,7 @@ import {
 } from '../../types';
 import {
   convertToElementValue,
+  createExecutionCode,
   elementValueChangedCodeParameters,
   fileToBase64,
   formatElementValue,
@@ -218,7 +221,7 @@ export const FormPanel: React.FC<Props> = ({
     }: {
       code: string;
       initial: unknown;
-      response?: FetchResponse | Response | null;
+      response?: FetchResponse | Response | DataQueryResponse | null;
       initialRequest?: () => Promise<void>;
       currentElements?: LocalFormElement[];
     }) => {
@@ -229,45 +232,10 @@ export const FormPanel: React.FC<Props> = ({
       /**
        * Function
        */
-      const f = new Function(
-        'options',
-        'data',
-        'response',
-        'elements',
-        'onChange',
-        'locationService',
-        'templateService',
-        'onOptionsChange',
-        'initialRequest',
-        'setInitial',
-        'json',
-        'initial',
-        'notifyError',
-        'notifySuccess',
-        'notifyWarning',
-        'toDataQueryResponse',
-        'context',
-        replaceVariables(code)
-      );
+      const f = createExecutionCode('context', replaceVariables(code));
 
       try {
         await f(
-          options,
-          data,
-          response,
-          currentElements || elementsRef.current,
-          onChangeElements,
-          locationService,
-          templateSrv,
-          onOptionsChange,
-          initialRequest,
-          setInitial,
-          initialRef.current,
-          initialRef.current,
-          notifyError,
-          notifySuccess,
-          notifyWarning,
-          toDataQueryResponse,
           requestCodeParameters.create({
             grafana: {
               locationService,
@@ -314,7 +282,6 @@ export const FormPanel: React.FC<Props> = ({
       templateSrv,
       onOptionsChange,
       setInitial,
-      initialRef,
       notifyError,
       notifySuccess,
       notifyWarning,
@@ -370,7 +337,7 @@ export const FormPanel: React.FC<Props> = ({
       return;
     }
 
-    let response: Response | FetchResponse | null;
+    let response: Response | FetchResponse | DataQueryResponse | null;
     let json: { [id: string]: unknown } = {};
 
     /**
@@ -390,7 +357,7 @@ export const FormPanel: React.FC<Props> = ({
       /**
        * Run Datasource Query
        */
-      const body = await getPayloadForRequest({
+      const payload = await getPayloadForRequest({
         request: {
           ...options.initial,
           payloadMode: PayloadMode.CUSTOM,
@@ -401,20 +368,20 @@ export const FormPanel: React.FC<Props> = ({
       });
 
       response = await datasourceRequest({
-        query: body,
+        query: options.initial.payload,
         datasource: options.initial.datasource,
         replaceVariables,
+        payload,
       }).catch((error: DataQueryError) => {
         setError(JSON.stringify(error));
         return null;
       });
 
-      if (response && response.ok) {
+      if (response && response.state === LoadingState.Done) {
         /**
          * Change Elements With Data Source Values
          */
-        const queryResponse = toDataQueryResponse(response as FetchResponse);
-        currentElements = getElementsWithFieldValues(queryResponse.data, RequestMethod.DATASOURCE);
+        currentElements = getElementsWithFieldValues(response.data, RequestMethod.DATASOURCE);
 
         /**
          * Update Elements and Initial Values
@@ -558,6 +525,8 @@ export const FormPanel: React.FC<Props> = ({
        */
       await executeCustomCode({ code: options.resetAction.code, initial: initialRef.current });
       setLoading(LoadingMode.NONE);
+
+      return;
     }
 
     /**
@@ -591,6 +560,7 @@ export const FormPanel: React.FC<Props> = ({
         highlight: false,
         highlightColor: '',
         confirm: false,
+        payload: options.resetAction.payload,
       },
       elements: elementsRef.current,
       initial: initialRef.current,
@@ -600,22 +570,22 @@ export const FormPanel: React.FC<Props> = ({
     /**
      * Datasource query
      */
-    const response: FetchResponse | null = await datasourceRequest({
-      query: payload,
+    const response = await datasourceRequest({
+      query: options.resetAction.payload,
       datasource: options.resetAction.datasource,
       replaceVariables,
+      payload,
     }).catch((error: DataQueryError) => {
       setError(JSON.stringify(error));
       return null;
     });
 
     let currentElements = elementsRef.current;
-    if (response && response.ok) {
+    if (response && response.state === LoadingState.Done) {
       /**
        * Change Elements With Data Source Values
        */
-      const queryResponse = toDataQueryResponse(response as FetchResponse);
-      currentElements = getElementsWithFieldValues(queryResponse.data, RequestMethod.DATASOURCE);
+      currentElements = getElementsWithFieldValues(response.data, RequestMethod.DATASOURCE);
 
       /**
        * Update Elements
@@ -640,6 +610,7 @@ export const FormPanel: React.FC<Props> = ({
     options.resetAction.datasource,
     options.resetAction.getPayload,
     options.resetAction.mode,
+    options.resetAction.payload,
     replaceVariables,
   ]);
 
@@ -683,16 +654,17 @@ export const FormPanel: React.FC<Props> = ({
     /**
      * Response
      */
-    let response: Response | FetchResponse | null;
+    let response: Response | FetchResponse | DataQueryResponse | null;
 
     /**
      * Datasource query
      */
     if (options.update.method === RequestMethod.DATASOURCE) {
       response = await datasourceRequest({
-        query: payload,
+        query: options.update.payload,
         datasource: options.update.datasource,
         replaceVariables,
+        payload,
       }).catch((error: DataQueryError) => {
         setError(JSON.stringify(error));
         return null;
@@ -829,7 +801,7 @@ export const FormPanel: React.FC<Props> = ({
    */
   const onElementValueChanged = useCallback(
     ({ elements, element }: { elements: LocalFormElement[]; element: LocalFormElement }) => {
-      const fn = new Function('context', replaceVariables(options.elementValueChanged));
+      const fn = createExecutionCode('context', replaceVariables(options.elementValueChanged));
 
       fn(
         elementValueChangedCodeParameters.create({
@@ -1111,7 +1083,10 @@ export const FormPanel: React.FC<Props> = ({
                     /**
                      * Skip not changed element
                      */
-                    if (element.value === initial[element.id]) {
+                    if (
+                      element.value === initial[element.id] &&
+                      options.confirmModal.elementDisplayMode !== ConfirmationElementDisplayMode.ALL
+                    ) {
                       return;
                     }