+++ src/utils/form-element.ts
@@ -22,6 +22,7 @@ import {
   LocalFormElement,
   ShowIfHelper,
 } from '../types';
+import { createExecutionCode } from './code';
 import { disableIfCodeParameters, getOptionsCodeParameters, showIfCodeParameters } from './code-parameters';
 import { getFieldValues } from './grafana';
 
@@ -238,14 +239,12 @@ export const isSectionCollisionExists = (sections: LayoutSection[], compareWith:
  */
 export const toLocalFormElement = (element: FormElement): LocalFormElement => {
   const showIf = element.showIf;
-
   let showIfFn: ShowIfHelper = () => true;
   if (showIf || showIf?.trim()) {
-    const fn = new Function('elements', 'replaceVariables', 'context', showIf);
+    const fn = createExecutionCode('context', showIf);
+
     showIfFn = ({ elements, replaceVariables }: { elements: FormElement[]; replaceVariables: InterpolateFunction }) =>
       fn(
-        elements,
-        replaceVariables,
         showIfCodeParameters.create({
           panel: {
             elements,
@@ -261,7 +260,8 @@ export const toLocalFormElement = (element: FormElement): LocalFormElement => {
 
   let disableIfFn: DisableIfHelper = () => false;
   if (disableIf || disableIf?.trim()) {
-    const fn = new Function('elements', 'replaceVariables', 'context', disableIf);
+    const fn = createExecutionCode('context', disableIf);
+
     disableIfFn = ({
       elements,
       replaceVariables,
@@ -270,8 +270,6 @@ export const toLocalFormElement = (element: FormElement): LocalFormElement => {
       replaceVariables: InterpolateFunction;
     }) =>
       fn(
-        elements,
-        replaceVariables,
         disableIfCodeParameters.create({
           panel: {
             elements,
@@ -316,7 +314,8 @@ export const toLocalFormElement = (element: FormElement): LocalFormElement => {
         }));
       };
     } else if (element.optionsSource === OptionsSource.CODE) {
-      const fn = new Function('data', 'elements', 'replaceVariables', 'context', element.getOptions || 'return []');
+      const fn = createExecutionCode('context', element.getOptions || 'return []');
+
       getOptions = ({
         data,
         elements,
@@ -327,9 +326,6 @@ export const toLocalFormElement = (element: FormElement): LocalFormElement => {
         replaceVariables: InterpolateFunction;
       }) =>
         fn(
-          data,
-          elements,
-          replaceVariables,
           getOptionsCodeParameters.create({
             panel: {
               data,